Fixed negative value display along with very large positive values. Changed data type for money account from UInt32 to Int64

+++ TS SE Tool/CustomClasses/Save/Items/PlayerData.cs
@@ -23,7 +23,7 @@ namespace TS_SE_Tool
         public uint ExperiencePoints { get; set; } = 0;
         public byte[] PlayerSkills { get; set; } = new byte[] { 0, 0, 0, 0, 0, 0 };
         //Bank
-        public uint AccountMoney { get; set; } = 0;
+        public long AccountMoney { get; set; } = 0;
         //Player
         public string HQcity { get; set; } = "";
         public string UserCompanyAssignedTruck { get; set; } = "";
